requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
mlxtend>=0.23.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
networkx>=3.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
